% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_kwic.R
\name{add_kwic}
\alias{add_kwic}
\title{add_kwic Function}
\usage{
add_kwic(corpus)
}
\arguments{
\item{corpus}{A data frame that contains a column for tokens}
}
\description{
This function creates columns 'word' for originally punctuated tokens, plus 'before' and 'after' for the surrounding context
}
\examples{
tier_exists("path/to/elan/file.eaf", "orth@NTP-M-1986")
}
\keyword{ELAN}
