% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_lang.R
\name{add_lang}
\alias{add_lang}
\title{add_lang Function}
\usage{
add_lang(file = "data/input/kpv_csys19570000-291_1a-Mezador.eaf",
  participant = "AXK-F-193X", linguistic_type = "wordT", lang = "kpv")
}
\arguments{
\item{corpus}{The corpus file}
}
\description{
This function adds a new tier into an ELAN file. It is prototypical at best.
}
\keyword{ELAN}

