% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_tier.R
\name{add_tier}
\alias{add_tier}
\title{add_tier Function}
\usage{

  add_tier(file = "/Volumes/langdoc/langs/kpv/kpv_izva20000321ChuprovII/kpv_izva20000321ChuprovII-backup.eaf",
  participant = "IIC-M-1937", linguistic_type = "var-sT",
  tier_id = "var-s", parent = "ref", time_alignable = TRUE,
  empty = FALSE, lang = "kpv")
}
\arguments{
\item{corpus}{The corpus file}
}
\description{
This function adds a new tier into an ELAN file. It is prototypical at best.
}
\keyword{ELAN}

