% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_stereotype.R
\name{change_stereotype}
\alias{change_stereotype}
\title{change_stereotype Function}
\usage{
change_stereotype(file, linguistic_type, stereotype, time_alignable,
  add = FALSE)
}
\arguments{
\item{file}{ELAN file}

\item{linguistic_type}{Name of the linguistic type}

\item{stereotype}{Name of the stereotype. Accepted values are, among some others, "Included_in", "Symbolic_Association" or "Symbolic_Subdivision".}

\item{time_alignable}{Here the accepted values are "true" or "false". Please note that these must be character vectors typed exactly like this.}

\item{add}{Should the type be added, or is it intended to modify an existing type with the same name?}
}
\description{
This function adds a new tier into an ELAN file. It is prototypical at best.
}
\keyword{ELAN}
