% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_words.R
\name{check_words}
\alias{check_words}
\title{check_words Function}
\usage{
check_words(eaf_file, corpus_rds)
}
\arguments{
\item{eaf_file}{Filename}

\item{corpus}{Data frame which contains corpus to compare}
}
\description{
This function takes one ELAN file reads the word tokens and compares those to the previously created data frame.
The idea is to find tokens which do not occur in corpus already and are possible candidates for being typos.
}
\examples{
check_words(eaf_file = "file.eaf", corpus_rds = "corpus.rds")
}
\keyword{ELAN}
