% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf2tcf.R
\name{eaf2tcf}
\alias{eaf2tcf}
\title{eaf2tcf}
\usage{

  eaf2tcf(eaf_file = "/Volumes/langdoc/langs/kpv//kpv_izva19661022-kotus13755_2az/kpv_izva19661022-kotus13755_2az.eaf")
}
\arguments{
\item{eaf_file}{ELAN file}
}
\description{
This function read an ELAN file and transforms it into TCF format, which can be used, for example, with Webanno
}
\keyword{ELAN}
