% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf2tsv.R
\name{eaf2tsv}
\alias{eaf2tsv}
\title{eaf2tsv}
\usage{

  eaf2tsv(eaf_file = "/Volumes/langdoc/langs/kpv/kpv_izva20140330-1FilippovaMV/kpv_izva20140330-1FilippovaMV-b/kpv_izva20140330-1FilippovaMV-b.eaf")
}
\arguments{
\item{eaf_file}{ELAN file}
}
\description{
This function read an ELAN file and transforms it into TCF format, which can be used, for example, with Webanno
}
\keyword{ELAN}
