% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf_align.R
\name{eaf_align}
\alias{eaf_align}
\title{eaf_align Function}
\usage{
eaf_align(corpus, c1 = "Before", c2 = "Token", c3 = "After", c4 = FALSE)
}
\arguments{
\item{corpus}{The corpus file}
}
\description{
This function launches a Shiny app. Please see the documentation in the address [github.com/langdoc/FRorpus-demo](http://github.com/langdoc/FRorpus-demo).
}
\keyword{ELAN}
