% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exc_token.R
\name{exc_token}
\alias{exc_token}
\title{exc_token Function}
\usage{
exc_token(corpus = corpus, regex = "test")
}
\arguments{
\item{corpus}{This is a data frame which contains column `Token`.}

\item{regex}{This is a regex one searches wants to exclude}
}
\description{
This function is equivalent to `filter(! grepl("regex", Token))`, but demands less typing. It is convenient to use this function in a dplyr pipe after the data frame output of read_eaf() function. It is very common that in order to exclude some tokens we can just define some regular expression patterns which do the job, after which we pass the results onward.
}
\examples{
corpus \%>\% find_token(".+ter$") \%>\% exc_token("^a.+")
}
\keyword{ELAN}

