% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_token.R
\name{find_token}
\alias{find_token}
\title{find_token Function}
\usage{
find_token(corpus = corpus, regex = "test")
}
\arguments{
\item{corpus}{This is a data frame which contains column `Token`.}

\item{regex}{This is a regex one searches with}
}
\description{
This function is equivalent to `filter(grepl("regex", Token))`, but demands less typing. It is convenient to use this function in a dplyr pipe after the data frame output of read_eaf() function. The results can be, as an example, piped into FRorpus() function in order to view the results in a Shiny application.
}
\examples{
find_token("^nd.+")
}
\keyword{ELAN}
