% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FRorpus.R
\name{FRorpus}
\alias{FRorpus}
\title{FRorpus Function}
\usage{
FRorpus(corpus, c1 = "Before", c2 = "Token", c3 = "After", c4 = FALSE)
}
\arguments{
\item{corpus}{The corpus file}
}
\description{
This function launches a Shiny app. Please see the documentation in the address [github.com/langdoc/FRorpus-demo](http://github.com/langdoc/FRorpus-demo).
}
\keyword{ELAN}

