% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_name.R
\name{get_session_name}
\alias{get_session_name}
\title{get_session_name Function extracts session name from the path. Super simple, but keeps that regex in one place}
\usage{
get_session_name(file_path)
}
\description{
This function just applies a regex
}
\examples{
get_session_name("/Volumes/data/kom/subfolder/another/good_session.eaf")
}
\keyword{ELAN}
