% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_notes.R
\name{harvest_notes}
\alias{harvest_notes}
\title{harvest_notes Function}
\usage{
harvest_notes(path = ".")
}
\arguments{
\item{path}{Where to look for files}
}
\description{
This function finds and parses note annotation files written in Emacs ld-notes-mode.
}
\keyword{documentation}
\keyword{language}
