% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_type_independent.R
\name{is_type_independent}
\alias{is_type_independent}
\title{is_tier_independent Function checks whether a linguistic type is independent or not}
\usage{
is_type_independent(eaf_file, linguistic_type)
}
\arguments{
\item{eaf_file}{This is a path to the file.}

\item{linguistic_type}{This is the linguistic type}
}
\description{
This function checks the linguistic type definition and returns information about whether it is possible to align the tier or not.
}
\examples{
is_tier_independent(path = "path/to/elan/file.eaf", tier = "refT")
}
\keyword{ELAN}
