% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_eaf.R
\name{log_eaf}
\alias{log_eaf}
\title{log_eaf Function}
\usage{
log_eaf(days = 3, path = ".", pattern = ".eaf$")
}
\arguments{
\item{days}{This defines how old files are returned. Default is 3 days.}

\item{path}{This is a path. Defaults to current working directory.}

\item{pattern}{This is a pattern in regex. Defaults has ELAN files ending with ".eaf".}
}
\description{
This function finds the newest ELAN files in a given directory.
}
\examples{
log_eaf(5)
}
\keyword{ELAN}
