% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mgsub.R
\name{mgsub}
\alias{mgsub}
\title{mgsub Function}
\usage{
mgsub(pattern, replacement, x, ...)
}
\arguments{
\item{The}{path to the text file you want to open}
}
\description{
This function applies multiple substitution patterns into gsub function
}

