% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cmdi.R
\name{read_cmdi}
\alias{read_cmdi}
\title{read_cmdi Function reads a CMDI file}
\usage{
read_cmdi(cmdi_file)
}
\arguments{
\item{cmdi_file}{The path to CMDI file}
}
\description{
read_cmdi Function reads a CMDI file
}
\examples{
read_cmdi('file.cmdi')
}
\keyword{ELAN}
