% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_elan_hash.R
\name{read_elan_hash}
\alias{read_elan_hash}
\title{read_elan_hash Function reads the identifier of the ELAN file}
\usage{
read_elan_hash(elan_file)
}
\arguments{
\item{elan_file}{ELAN file}
}
\description{
This function adds a new tier into an ELAN file. It is prototypical at best.
}
\keyword{ELAN}
