% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kpv.R
\name{read_kpv}
\alias{read_kpv}
\title{read_kpv Function}
\usage{

  read_kpv(file = "/Volumes/langdoc/langs/kpv/kpv_udo20120330SazinaJS-dream/kpv_udo20120330SazinaJS-dream.eaf",
  SS_tier = "wordT", SA_tier = "orthT", DEF_tier = "refT",
  simplify = TRUE)
}
\arguments{
\item{file}{The path to ELAN file which we want to parse}

\item{SS_tier}{Linguistic type of the Symbolic Subdivision tier (usually contains the tokenized wordforms)}

\item{SA_tier}{Linguistic type of the Symbolic Association tier}

\item{DEF_tier}{Linguistic type of the independent tier}
}
\description{
This is an updated version of the old function, which has been renamed to read_eaf_old(). Instead of trying to do everything this function is supposedly used with `plyr` in order to parse multiple files. The function parses ELAN files. Dummy tokens that contain no information are erased automatically. The files that are for some reason not parsable are skipped. In actual use the biggest problems are connected to structural irregularity of ELAN files in corpus. It is used ideally in connection with scripts that are able to parse IMDI or CMDI files. Please use `log_eaf()` function to see which files have been changed recently, those are usually the ones containing problems.
}
\examples{
read_kpv(path = "corpora/kpv/session_1.eaf", DEF_tier = "refT", SA_tier = "orthT", SS_tier = "wordT")
}
\keyword{ELAN}
