% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lang.R
\name{read_lang}
\alias{read_lang}
\title{read_lang Function reads tier language attributes from a single ELAN files.}
\usage{

  read_lang(eaf_file = "/Volumes/langdoc/langs/kpv/kpv_izva20140404IgusevJA/kpv_izva20140404IgusevJA.eaf",
  linguistic_type = "orthT", def_lang = "kpv")
}
\arguments{
\item{eaf_file}{This is a path to the file.}

\item{tier}{This is the linguistic type of the tier}
}
\description{
This function reads just an individual tier of one ELAN file
}
\examples{
read_tier(path = "corpora/kpv/", tier = "wordT")
}
\keyword{ELAN}
