% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_tier.R
\name{read_tier}
\alias{read_tier}
\title{read_tier Function read individual tiers from single ELAN files.}
\usage{
read_tier(eaf_file, tier = "wordT", independent = FALSE)
}
\arguments{
\item{eaf_file}{This is a path to the file.}

\item{tier}{This is the linguistic type of the tier}
}
\description{
This function reads just an individual tier of one ELAN file
}
\examples{
read_tier(path = "corpora/kpv/", tier = "wordT")
}
\keyword{ELAN}

