% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_timeslots.R
\name{read_timeslots}
\alias{read_timeslots}
\title{read_timeslots Function grabs the time slot information from ELAN file}
\usage{
read_timeslots(eaf_file, read_file = T, xml_object = F)
}
\arguments{
\item{eaf_file}{This is a path to the file.}
}
\description{
This function reads the information about time slots
}
\examples{
read_timeslots(eaf_file = "path/to/elan/file.eaf")
}
\keyword{ELAN}
