% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_tier.R
\name{remove_tier}
\alias{remove_tier}
\title{remove_tier Function}
\usage{
remove_tier(file, linguistic_type = FALSE, tier_id = FALSE)
}
\arguments{
\item{file}{ELAN file to be parsed}

\item{linguistic_type}{The linguistic type which is to be removed. If tier_id is used, this should be empty.}

\item{tier_id}{The beginning of the tier name which should be removed. The idea is that one doesn't need to define the whole tier name, which may contain speaker id's and other information. This is especially suitable strategy when dealing with files which don't have well differentiated linguistic types. If linguistic_type is defined, this should be left empty.}
}
\description{
This function adds a new tier into an ELAN file. It is prototypical at best.
}
\keyword{ELAN}

