% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_tier.R
\name{rename_tier}
\alias{rename_tier}
\title{rename_tier Function renames a tier}
\usage{
rename_tier(filename, current_name, new_name, suffix)
}
\arguments{
\item{filename}{ELAN file}

\item{current_name}{Name of the linguistic type}

\item{new_name}{Name to be set}

\item{suffix}{Appends a suffix into saved filename}
}
\description{
This function adds a new tier into an ELAN file. It is prototypical at best.
}
\keyword{ELAN}
