% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_type.R
\name{rename_type}
\alias{rename_type}
\title{rename_type Function Renames a linguistic type}
\usage{
rename_type(filename, current_name, new_name, suffix)
}
\arguments{
\item{eaf_file}{This is a path to the file.}
}
\description{
This function rewrites the content of reference tier in an ELAN file
}
\examples{
rename_type(eaf_file = "./corpus/elan_file.eaf", current_name = 'refT', new_name = 'referenceT', suffix = '-test')
}
\keyword{ELAN}
