% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_eaf.R
\name{test_eaf}
\alias{test_eaf}
\title{test_eaf Function}
\usage{
test_eaf(path = ".", pattern = "\\\\.eaf$", tier = "wordT",
  recursive = TRUE)
}
\arguments{
\item{path}{This is a path. Defaults to current directory.}

\item{pattern}{This is a pattern in regex. Defaults has elan files only.}

\item{tier}{This is the linguistic type of the tier}

\item{recursive}{Should the function search to all subfolders on the path? Default = FALSE.}
}
\description{
This function tests different tiers
}
\examples{
test_eaf(path = "corpora/kpv/", tier = "refT", recursive = TRUE)
}
\keyword{ELAN}

