% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tier_exists.R
\name{tier_exists}
\alias{tier_exists}
\title{tier_exists Function}
\usage{
tier_exists(eaf_file, tier)
}
\arguments{
\item{eaf_file}{The path to ELAN file which we want to parse}

\item{tier}{Tier name}
}
\description{
This function simply tests whether ELAN file has a tier or not
}
\examples{
tier_exists("path/to/elan/file.eaf", "orth@NTP-M-1986")
}
\keyword{ELAN}
