% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tier_has_content.R
\name{tier_has_content}
\alias{tier_has_content}
\title{tier_has_content Function}
\usage{
tier_has_content(eaf_file, tier, value = F)
}
\arguments{
\item{eaf_file}{The path to ELAN file which we want to parse}

\item{tier}{Tier name}

\item{value}{Whether the function should return the number of items on the tier instead of a boolean value}
}
\description{
This function simply tests whether ELAN file has a tier or not
}
\examples{
tier_has_content("path/to/elan/file.eaf", "orth@NTP-M-1986")
}
\keyword{ELAN}
