% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_eaf.R
\name{tokenize_eaf}
\alias{tokenize_eaf}
\title{tokenize_eaf Function}
\usage{
tokenize_eaf(file = "test/sjt19890000voronova1989a-05.eaf",
  participant = "OVV", linguistic_type = "orthT", lang = "kpv")
}
\arguments{
\item{corpus}{The corpus file}
}
\description{
This function tokenizes an ELAN file. It takes the content of the orthT tier and tokenizes that onto wordT tier.
}
\keyword{ELAN}
