% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transliterate.R
\name{transliterate}
\alias{transliterate}
\title{transliterate Function}
\usage{
transliterate(data, model)
}
\arguments{
\item{data}{Data into which the function is applied}

\item{model}{Path to the text file which contains the model for transliteration}
}
\description{
This functions converts character vectors from one character set to another.
}
\examples{
transliterate("тест", "kpv-cyr2ipa")
}
\keyword{ELAN}
\keyword{linguistics}
