% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_references.R
\name{write_references}
\alias{write_references}
\title{write_references Function read individual tiers from single ELAN files.}
\usage{

  write_references(eaf_file = "~/langdoc/kpv//kpv_izva20140331-1burki/kpv_izva20140331-1burki.eaf")
}
\arguments{
\item{eaf_file}{This is a path to the file.}
}
\description{
This function rewrites the content of reference tier in an ELAN file
}
\examples{
write_references(eaf_file = "./corpus/elan_file.eaf")
}
\keyword{ELAN}
