% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_segments.R
\name{write_segments}
\alias{write_segments}
\title{write_segments Function}
\usage{

  write_segments(eaf_file = "/Volumes/langdoc/langs/kpv/kpv_izva20140330-1FilippovaMV/kpv_izva20140330-1FilippovaMV-b/kpv_izva20140330-1FilippovaMV-b.eaf",
  linguistic_type = "refT")
}
\arguments{
\item{eaf_file}{ELAN file}

\item{tier_prefix}{Tiername prefix that specifies wanted items}
}
\description{
This function takes the ELAN file and writes the found time codes into `.segments` format used by pyAudioAnalysis
}
\keyword{ELAN}
\keyword{langdoc}
\keyword{linguistics}
