% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lang.R
\name{add_lang}
\alias{add_lang}
\title{add_lang Function}
\usage{
add_lang(eaf_file = "testfile.eaf", participant = "AXK-F-193X",
  linguistic_type = "wordT", lang = "kpv")
}
\arguments{
\item{eaf_file}{Filename}

\item{participant}{The id of the participant}

\item{lang}{ISO-code for the language which has to be added}

\item{linguistic}{type The linguistic type of the tier}
}
\description{
This function adds a language reference into ELAN tiers which are specified by type and speaker
}
\keyword{ELAN}
