% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exc_token.R
\name{exc_token}
\alias{exc_token}
\title{exc_token Function}
\usage{
exc_token(corpus = corpus, regex = "чолӧм")
}
\arguments{
\item{regex}{This is a regex one searches wants to exclude}
}
\description{
This function is equivalent to `filter(! grepl("regex", Token))`
}
\examples{
find_token(".+лаздор.?")
}
\keyword{ELAN}
\keyword{linguistics}

