% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_token.R
\name{find_token}
\alias{find_token}
\title{find_token Function}
\usage{
find_token(corpus = corpus, regex = "чолӧм")
}
\arguments{
\item{regex}{This is a regex one searches with}
}
\description{
This function is equivalent to `filter(grepl("regex", Token))`
}
\examples{
find_token(".+лаздор.?")
}
\keyword{ELAN}
\keyword{linguistics}

