% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRorpus.R
\name{FRorpus}
\alias{FRorpus}
\title{FRorpus Function}
\usage{
FRorpus(data, c1 = "before", c2 = "token", c3 = "after", c4 = FALSE)
}
\arguments{
\item{corpus}{The corpus file}
}
\description{
This function launches a Shiny app. Please see the documentation in the address [github.com/langdoc/FRorpus-demo](http://github.com/langdoc/FRorpus-demo).
}
\keyword{ELAN}
