% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/open_eaf.R
\name{open_eaf}
\alias{open_eaf}
\title{open_eaf Function}
\usage{
open_eaf(corpus = corpus_kpv, row = 1, program = FALSE)
}
\arguments{
\item{corpus}{To which data frame you want to apply the function}

\item{row}{The number of dataframe row you want to open}

\item{program}{The program with which you want to open ELAN file}
}
\description{
This function opens an ELAN file from R. Intended use covers mainly situations where another function has, as an example, listed all files which have incorrectly linked media files which user wants to relocate file by file. In principle one could also develop it to open file in a specific location, but as far as I see this involves manipulating .psfx file which defines the selection on the timeline when opened.
}
\examples{
open_eaf(5)
}
\keyword{ELAN}

